extract project resource schema to resource_schema.go
